backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
python-dotenv
openpyxl
psycopg2-binary
python-jose[cryptography]
bcrypt
